hopr_transport

Module protocol

Transport binary protocol layer (codec, pipeline, heartbeat, stream). Collection of objects and functionality allowing building of p2p or stream protocols for the higher business logic layers.

Contents

Supported protocol configurations:

mix

ack

heartbeat

Re-exports

pub use counters::PeerProtocolCounterRegistry;

pub use counters::PeerProtocolCounters;

Modules

codec 🔒

Coder and decoder for the transport binary protocol layer

counters

Per-peer protocol conformance counters.

errors

Errors produced by the crate.

heartbeat

heartbeat p2p protocol

pipeline 🔒

Packet pipeline for the HOPR protocol. HOPR packet processing pipeline.

stream

Stream processing utilities Infrastructure supporting converting a collection of PeerId split libp2p_stream managed individual peer-to-peer libp2p::swarm::Streams.

Structs

AcknowledgementPipelineConfig

Configuration for the acknowledgement processing pipeline.

PacketPipelineBuilder

Builder for constructing the HOPR packet pipeline for a specific node type.

PacketPipelineConfig

Overall configuration of the input/output packet processing pipeline.

Unset

Placeholder type used by PacketPipelineBuilder for generic parameters that have not yet been provided via the corresponding builder method.

Enums

NodeType

Node type for which the packet processing pipeline is being constructed.

PacketPipelineProcesses

Constants

CURRENT_HOPR_MSG_PROTOCOL

HOPR_PACKET_SIZE 🔒

Type Aliases

HoprBinaryCodec