hopr_transport
Type Alias ServiceId

pub type ServiceId = u32;

Describes a node service target. These are specialized SessionTargets that are local to the Exit node and have different purposes, such as Cover Traffic.

These targets cannot be sealed from the Entry node.