Crate hopr_transport

The crate aggregates and composes individual transport level objects and functionality into a unified crate::HoprTransport object with the goal of isolating the transport layer and defining a fully specified transport API.

See also the hopr_protocol_start crate for details on Start sub-protocol which initiates a Session.

As such, the transport layer components should be only those that are directly necessary to:

send and receive a packet, acknowledgement or ticket aggregation request

send and receive a network telemetry request

automate transport level processes

algorithms associated with the transport layer operational management

interface specifications to allow modular behavioral extensions

Re-exports

pub use stats::PeerPacketStats;

pub use crate::config::HoprProtocolConfig;

pub use crate::stats::PeerPacketStatsSnapshot;

pub use hopr_transport_session as session;

pub use hopr_api as api;

Modules

config

Configuration of the crate::HoprTransport.

constants

Constants used and exposed by the crate.

errors

Errors used by the crate.

socket

stats

Structs

ApplicationData

Represents the to-be-sent or received decrypted packet carrying the application-layer data.

ApplicationDataIn

Wrapper for incoming ApplicationData with optional [IncomingPacketInfo].

ApplicationDataOut

Wrapper for outgoing ApplicationData with optional [IncomingPacketInfo].

ChainKeypair

Represents a keypair consisting of a secp256k1 private and public key

ChannelTicketStatistics

Contains ticket statistics for one or more channels.

HalfKeyChallenge

Represents a challenge for the half-key in Proof of Relay.

HoprSession

Represents the Session protocol socket over HOPR.

HoprTransport

Interface into the physical transport mechanism allowing all off-chain HOPR-related tasks on the transport.

IncomingSession

Wrapper for incoming HoprSession along with other information extracted from the Start protocol during the session establishment.

Multiaddr

Representation of a Multiaddr.

Observations

Observations related to a specific peer in the network.

OffchainKeypair

Represents a keypair consisting of an Ed25519 private and public key

OffchainPublicKey

Represents an Ed25519 public key.

PeerId

Identifier of a peer of the network.

PingQueryReplier

Helper object allowing to send a ping query as a wrapped channel combination that can be filled up on the transport part and awaited locally by the Pinger.

SessionClientConfig

Configuration for the session.

SessionId

Unique ID of a specific Session in a certain direction.

SessionLifetimeSnapshot

Snapshot of session lifetime metrics.

SessionStatsSnapshot

Complete snapshot of all session metrics at a point in time.

SurbBalancerConfig

Configuration for the SurbBalancer.

Enums

Health

Network health represented with colors, where green is the best and red is the worst possible observed network quality.

HoprTransportProcess

PeerDiscovery

Processed indexer generated events.

ProbeError

Protocol

Protocol describes all possible multiaddress protocols.

RoutingOptions

Represents routing options in a mixnet with a maximum number of hops.

SessionAckMode

The acknowledgement mode configured for the session.

SessionCapability

Individual capabilities of a Session.

SessionLifecycleState

The lifecycle state of a session from the perspective of metrics.

SessionManagerError

SessionTarget

Defines what should happen with the data at the recipient where the data from the established session are supposed to be forwarded to some target.

Tag

Tags distinguishing different application-layer protocols.

TicketEvent

Ticket events emitted from the packet processing pipeline.

TransportSessionError

Enumeration of errors thrown from this library.

Constants

APPLICATION_TAG_RANGE

SESSION_MTU

Number of bytes that can be sent in a single Session protocol payload.

SURB_SIZE

Size of the HOPR SURB in bytes.

Traits

Keypair

Represents a generic key pair The keypair contains a private key and public key. Must be comparable in constant time and zeroized on drop.

NetworkView

Trait representing a read-only view of the network state.

Observable

Functions

transfer_session

Convenience function to copy data in both directions between a Session and arbitrary async IO stream. This function is only available with Tokio and will panic with other runtimes.

Type Aliases

Hash

Represents an Ethereum 256-bit hash value.

HoprPseudonym

Alias for the Pseudonym used in the HOPR protocol.

ServiceId

Describes a node service target. These are specialized SessionTargets that are local to the Exit node and have different purposes, such as Cover Traffic.

SessionCapabilities

Set of Session capabilities.