hopr_transport
Module config

Configuration of the crate::HoprTransport.

Structs

HeartbeatConfig

Configuration for the Heartbeat mechanism

HostConfig

Configuration of the listening host.

NetworkConfig

Configuration for the crate::network::Network object

ProtocolConfig

Configuration of the P2P protocols.

TransportConfig

Configuration of the physical transport mechanism.

Enums

HostType

Enumeration of possible host types.

Functions

is_reachable_domain

Check whether the string is an actual reachable domain.

looks_like_domain

Check whether the string looks like a valid domain.

validate_external_host

Validates the HostConfig to be used as an external host