hopr_transport

Module config

Configuration of the crate::HoprTransport.

Structs

HoprCodecConfig

Configuration of [HoprEncoder] and [HoprDecoder].

HoprPacketPipelineConfig

Configuration of the HOPR packet pipeline.

HoprProtocolConfig

Complete configuration of the HOPR protocol stack.

HoprTicketProcessorConfig

Configuration for the HOPR ticket processor within the packet pipeline.

HostConfig

Configuration of the listening host.

ProbeConfig

Configuration for the probing mechanism

SessionGlobalConfig

Global configuration of Sessions and the Session manager.

SurbStoreConfig

Configuration for the SURB cache.

TransportConfig

Configuration of the physical transport mechanism.

Enums

HostType

Enumeration of possible host types.

Functions

is_reachable_domain

Check whether the string is an actual reachable domain.

looks_like_domain

Check whether the string looks like a valid domain.