hopr_transport
Function execute_on_tick

pub async fn execute_on_tick<F>(cycle: Duration, action: impl Fn() -> F)
where
    F: Future<Output = ()> + Send,

Construct an infinitely running background loop producing ticks with a given period with the maximum tick duration at most the period.